refactor pm quality ai summary route
