+++ src/app/router_registry.py
@@ -51,6 +51,9 @@ from app.routers.dpm_command_center_outcome_reviews import (
 from app.routers.dpm_command_center_pm_quality import (
     router as dpm_command_center_pm_quality_router,
 )
+from app.routers.dpm_command_center_pm_quality_ai import (
+    router as dpm_command_center_pm_quality_ai_router,
+)
 from app.routers.dpm_command_center_pm_quality_fairness import (
     router as dpm_command_center_pm_quality_fairness_router,
 )
@@ -211,6 +214,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_outcome_review_narratives_router)
     app.include_router(dpm_command_center_outcome_review_lookups_router)
     app.include_router(dpm_command_center_pm_quality_router)
+    app.include_router(dpm_command_center_pm_quality_ai_router)
     app.include_router(dpm_command_center_pm_quality_fairness_router)
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
